requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
google-generativeai==0.8.3  # Pastikan untuk mengecek versi terbaru yang sesuai
